requirements.txt
APScheduler==3.6.3
autopep8==1.5.7
certifi==2020.12.5
pycodestyle==2.7.0
python-dotenv==0.17.1
python-telegram-bot==13.5
pytz==2021.1
six==1.16.0
toml==0.10.2
tornado==6.1
tzlocal==2.1
